requirements.txt
-i https://pypi.org/simple
asgiref==3.5.2 ; python_version >= '3.7'
certifi==2022.9.24 ; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dj-database-url==1.0.0
dj-rest-auth==2.2.5
django==4.1.2
django-allauth==0.51.0
django-cors-headers==3.13.0
django-on-heroku==1.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
drf-yasg==1.21.4
gunicorn==20.1.0
idna==3.4 ; python_version >= '3.5'
inflection==0.5.1 ; python_version >= '3.5'
itypes==1.2.0
jinja2==3.1.2 ; python_version >= '3.7'
markupsafe==2.1.1 ; python_version >= '3.7'
oauthlib==3.2.1 ; python_version >= '3.6'
packaging==21.3 ; python_version >= '3.6'
pillow==9.2.0
psycopg2-binary==2.9.4
pycparser==2.21
pygments==2.13.0
pyjwt[crypto]==2.5.0 ; python_version >= '3.7'
pyparsing==3.0.9 ; python_full_version >= '3.6.8'
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.4
requests==2.28.1 ; python_version >= '3.7' and python_version < '4'
requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
ruamel.yaml==0.17.21 ; python_version >= '3'
ruamel.yaml.clib==0.2.6 ; platform_python_implementation == 'CPython' and python_version < '3.11'
setuptools==65.5.0 ; python_version >= '3.7'
sqlparse==0.4.3 ; python_version >= '3.5'
types-cryptography==3.3.23
uritemplate==4.1.1 ; python_version >= '3.6'
urllib3==1.26.12 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
whitenoise==6.2.0
